renamed gitStatusOutput file. added unstaged and staged files. type of change in parentheses after filename (likely needs further editing).
